+++ packages/svelte-query/tests/TestComponents/SubscriptionWrapper.svelte
@@ -1,0 +1,9 @@
+<script>
+	import { QueryClientProvider } from '@tanstack/svelte-query'
+	import SubscriptionComponent from './SubscriptionComponent.svelte';
+	export let queryClient, subscriptionCreator
+</script>
+
+<QueryClientProvider client={queryClient}>
+	<SubscriptionComponent subscriptionCreator={subscriptionCreator}/>
+</QueryClientProvider>