+++ packages/golang-client/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.2](https://github.com/wundergraph/wundergraph/compare/@wundergraph/golang-client@0.5.1...@wundergraph/golang-client@0.5.2) (2023-03-30)
+
+**Note:** Version bump only for package @wundergraph/golang-client
+
 ## [0.5.1](https://github.com/wundergraph/wundergraph/compare/@wundergraph/golang-client@0.5.0...@wundergraph/golang-client@0.5.1) (2023-03-29)
 
 **Note:** Version bump only for package @wundergraph/golang-client