+++ packages/swr/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.14.2](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.14.1...@wundergraph/swr@0.14.2) (2023-03-30)
+
+**Note:** Version bump only for package @wundergraph/swr
+
 ## [0.14.1](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.14.0...@wundergraph/swr@0.14.1) (2023-03-29)
 
 **Note:** Version bump only for package @wundergraph/swr